diversify error handling for object parsing
