+++ include/parser.h
@@ -10,8 +10,7 @@ char		*stra_iteri(char **arr, char *set, int id);
 double		get_ratio(char *str, t_errors *exit_code);
 t_vector	get_color(char *str, t_errors *exit_code);
 t_vector	get_vector(char *str, t_errors *exit_code, int pos);
-t_object	*get_objects(char **file, size_t *size, char *set,
-				t_object (parse)(char *, t_errors *));
+t_errors	get_objects(t_object **obj, char **file, char *set, size_t *size);
 t_object	parse_sphere(char *str, t_errors *ecode);
 t_object	parse_cylinder(char *str, t_errors *ecode);
 t_object	parse_plane(char *str, t_errors *ecode);
@@ -20,6 +19,8 @@ t_errors	set_camera(char **file, t_map *map);
 t_errors	get_obj_arr(char **file, t_map *map);
 t_errors	set_lighting(char **file, t_map *map);
 size_t		get_size(char **file, char *set);
+t_object	*realloc_arr(size_t size, t_object *src);
+t_object	parse(char *str, char *set, t_errors *ecode);
 int			ft_iswhitespcs(char c);
 
 #endif
