+++ src/parser/get_utils.c
@@ -21,49 +21,6 @@ size_t	get_size(char **file, char *set)
 	return (size);
 }
 
-t_object	parse_sphere(char *str, t_errors *ecode)
-{
-	char		**args;
-	t_object	sphere;
-
-	args = ft_split(str, '\t');
-	if (ft_stra_len(args) != 4)
-		return (*ecode = 1, new_sphere(new_vec(0, 0, 0), 0, new_vec(0, 0, 0)));
-	sphere = new_sphere(get_vector(args[1], ecode, 1),
-			get_ratio(args[2], ecode), get_color(args[3], ecode));
-	return (sphere);
-}
-
-t_object	parse_cylinder(char *str, t_errors *ecode)
-{
-	char		**args;
-	double		size[2];
-	t_object	cylinder;
-
-	args = ft_split(str, '\t');
-	if (ft_stra_len(args) != 6)
-		return (*ecode = 1, new_cylinder(new_vec(0, 0, 0), new_vec(0, 0, 0),
-				size, new_vec(0, 0, 0)));
-	size[0] = get_ratio(args[3], ecode);
-	size[1] = get_ratio(args[4], ecode);
-	cylinder = new_cylinder(get_vector(args[1], ecode, 1),
-			get_vector(args[2], ecode, 0), size, get_color(args[5], ecode));
-	return (cylinder);
-}
-
-t_object	parse_plane(char *str, t_errors *ecode)
-{
-	char		**args;
-	t_object	plane;
-
-	args = ft_split(str, '\t');
-	if (ft_stra_len(args) != 4)
-		return (*ecode = 1, new_plane(new_vec(0, 0, 0),
-				new_vec(0, 0, 0), new_vec(0, 0, 0)));
-	plane = new_plane(get_vector(args[1], ecode, 1),
-			get_vector(args[2], ecode, 0), get_color(args[3], ecode));
-	return (plane);
-}
 
 t_object	*join_objs(t_object *dst, t_object *src, size_t prev, size_t len)
 {
@@ -91,4 +48,33 @@ t_object	*join_objs(t_object *dst, t_object *src, size_t prev, size_t len)
 	free(dst);
 	free(src);
 	return (res);
+}
+
+t_object	*realloc_arr(size_t size, t_object *src)
+{
+	size_t		i;
+	t_object	*res;
+
+	i = 0;
+	if (src == NULL)
+		return (NULL);
+	res = malloc(size * sizeof(t_object));
+	while (i < size)
+	{
+		res[i] = src[i];
+		i++;
+	}
+	free(src);
+	return (res);
+}
+
+t_object	parse(char *str, char *set, t_errors *ecode)
+{
+	if (!ft_strncmp(set, "sp", 3))
+		return (parse_sphere(str, ecode));
+	if (!ft_strncmp(set, "cy", 3))
+		parse_cylinder(str, ecode);
+	if (!ft_strncmp(set, "pl", 3))
+		parse_plane(str, ecode);
+	return (parse_sphere(str, ecode));
 }
